Iteration 4
Late but looking for points please
Sprint Review
This iteration was important to prevent hard to detect errors in this
app.  The code was pretty simple and can be found in the CourseGateway
class, the Spec entry class. The IncListActivity also has an assert
statement.   The new code has been uploaded to GitHub for your review.